roomsplit/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roomsplit-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "sqlalchemy[asyncio]==2.0.36",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2",
    "python-multipart==0.0.20",
    "redis==5.2.1",
    "celery==5.4.0",
    "qrcode[pil]==8.0.0",
    "twilio==9.4.1",
    "httpx==0.28.1",
    "python-dotenv==1.0.1",
    "Pillow==11.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
